+++ NpgsqlRestTests/TsClientTests/DoNothingTests.cs
@@ -73,6 +73,9 @@ export async function tsclientTestDoNothing() : Promise<void> {
         private const string ExpectedStatus = """
 const baseUrl = "";
 
+type ApiError = {status: number; title: string; detail?: string | null};
+type ApiResult<T> = {status: number, response: T, error: ApiError | undefined};
+
 
 /**
 * function tsclient_test.do_nothing_status()
@@ -82,17 +85,17 @@ const baseUrl = "";
 * comment on function tsclient_test.do_nothing_status is 'tsclient_module=do_nothing_status
 * tsclient_status_code=true';
 * 
-* @returns {status: number, error: {status: number; title: string; detail?: string | null} | undefined}
+* @returns {status: number, error: ApiError | undefined}
 * 
 * @see FUNCTION tsclient_test.do_nothing_status
 */
-export async function tsclientTestDoNothingStatus() : Promise<{status: number, error: {status: number; title: string; detail?: string | null} | undefined}> {
+export async function tsclientTestDoNothingStatus() : Promise<{status: number, error: ApiError | undefined}> {
     const response = await fetch(baseUrl + "/api/tsclient-test/do-nothing-status", {
         method: "POST",
     });
     return {
         status: response.status,
-        error: !response.ok && response.headers.get("content-length") !== "0" ? await response.json() as {status: number; title: string; detail?: string | null} : undefined
+        error: !response.ok && response.headers.get("content-length") !== "0" ? await response.json() as ApiError : undefined
     };
 }